(BSR) : move GeolocationBanner into shared files
